% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{decode}
\alias{decode}
\alias{decode_bbox}
\title{Decode a geohash}
\usage{
decode(geohash)

decode_bbox(geohash)
}
\arguments{
\item{geohash}{a character vector of geohash codes}
}
\value{
\itemize{
\item \code{decode()} returns a \code{data.frame} with four columns: \code{x}, \code{y}, and \code{x_error}, \code{y_error}
\item \code{decode_bbox()} returns a list of \code{sf} \code{bbox} objects
}
}
\description{
Decodes a vector of geohashes.
}
\examples{
decode("eyywe2zq")
decode_bbox("eyywe2zq")
}
