% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{encode}
\alias{encode}
\title{Encode a coordinate to a geohash}
\usage{
encode(x, y, length)
}
\arguments{
\item{x}{a numeric vector of longitudes. Must be within the range of \link{-180, 180} otherwise an \code{NA} will be returned.}

\item{y}{a numeric vector of latitudes. Must be within the range of \link{-90, 90} otherwise an \code{NA} will be returned.}

\item{length}{a scalar integer between the values of 1 and 12.}
}
\description{
Given a vector of x and y coordinates, returns the geohash of the location.
Coordinates must be provided in longitude and latitude. In the case that an invalid
longitude or latitude value is provided, an \code{NA} is returned and not an error.
}
