% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash-package.R
\docType{package}
\name{geohash-package}
\alias{geohash}
\alias{geohash-package}
\title{geohash: Functions for geohash encoding and decoding}
\description{
A way to use a set of wrappers around 'rgeohash' rust crate. It includes encoding and decoding functions and a recipe step_geocode for geohash encoding.
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
