% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash.R
\docType{package}
\name{geohash}
\alias{geohash}
\alias{geohash-package}
\title{Tools for handling URLs}
\description{
The geohash package provides tools to encode lat/long pairs into geohashes, decode those geohashes,
and identify neighbours their neighbours.
}
\seealso{
the \href{https://cran.r-project.org/package=geohash/vignettes/geohash.html}{package vignette}.
}

