% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohash.R
\name{geohash_decode.default}
\alias{geohash_decode.default}
\title{Decode a Geohashes to a dataframe containing latitude and longitude}
\usage{
\method{geohash_decode}{default}(geohash)
}
\arguments{
\item{geohash}{Geohash string}
}
\value{
Data.frame with latitude, longitude, and their respective error values
}
\description{
Decode a Geohashes to a dataframe containing latitude and longitude
}
\examples{
geohash_decode("u1hzz6")
geohash_decode("u1m2fw")
geohash_decode("u1m2fwtr49m2fwtr49m2")
}

