% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohash.R
\name{geohash_decode.list}
\alias{geohash_decode.list}
\title{Decode lists containing Geohashes to a dataframe containing latitude and longitude}
\usage{
\method{geohash_decode}{list}(li)
}
\arguments{
\item{li}{List of Geohash strings}
}
\value{
Data.frame with latitude, longitude, and their respective error values
}
\description{
Decode lists containing Geohashes to a dataframe containing latitude and longitude
}
\examples{
geohash_decode(list("u1m2fw", "u1m2fwtr49m2fwtr49m2"))
}

