% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohash.R
\name{geohash_encode.data.frame}
\alias{geohash_encode.data.frame}
\title{Encode dataframes containing latitude and longitude to a dataframe containing Geohashes}
\usage{
\method{geohash_encode}{data.frame}(df, precision = 12)
}
\arguments{
\item{df}{A data frame where you have specified coordinates in the latitude and longitude column}

\item{precision}{Precision/length of the resulting Geohash}
}
\value{
Data.frame with the Geohash strings
}
\description{
Encode dataframes containing latitude and longitude to a dataframe containing Geohashes
}
\examples{
lats=c(52,47,-18)
lons=c(7,20,0)
df = data.frame(latitude=lats, longitude=lons)
geohash_encode(df)
}

