% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohash.R
\name{geohash_encode.default}
\alias{geohash_encode.default}
\title{Encode latitude and longitude to a Geohash}
\usage{
\method{geohash_encode}{default}(latitude, longitude, precision = 12)
}
\arguments{
\item{latitude}{Coordinate latitude}

\item{longitude}{Coordinate longitude}

\item{precision}{Precision/length of the resulting Geohash}
}
\value{
The Geohash string
}
\description{
Encode latitude and longitude to a Geohash
}
\examples{
geohash_encode(52, 7, precision=6)
geohash_encode(52.54123, 7.67123)
}

