\name{geohash-package}
\alias{geohash-package}
\alias{geohash}
\docType{package}
\title{
Convert coordinates between latitude/longitudes format and geohash
}
\description{
Convert coordinates between latitude/longitudes format and geohash
}
\details{
\tabular{ll}{
Package: \tab geohash\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-08-02\cr
License: \tab BSD 2-clause\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Christoph Kisfeld

Maintainer: Christoph Kisfeld <christoph.kisfeld@gmail.com>
}
\references{
https://en.wikipedia.org/wiki/Geohash
http://geohash.org/
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
