% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_geohash.R
\name{step_geohash}
\alias{step_geohash}
\title{Recipe to geohash encode longitude and latitude}
\usage{
step_geohash(
  recipe,
  lon = NULL,
  lat = NULL,
  role = NA,
  trained = FALSE,
  options = list(length = 8, parallel = FALSE),
  name = "geohash",
  columns = NULL,
  skip = FALSE,
  id = rand_id("geohash")
)
}
\arguments{
\item{recipe}{recipe}

\item{lon}{longitude}

\item{lat}{latitude}

\item{role}{role}

\item{trained}{FALSE}

\item{options}{list(length = 8, parallel = FALSE)}

\item{name}{"geohash"}

\item{columns}{NULL}

\item{skip}{FALSE}

\item{id}{rand_id("geohash")}
}
\description{
Recipe to geohash encode longitude and latitude
}
